Applying Move Method tecnique of Moving Characteristics Between Objects in getCommands(), createCommand() of ParamModel class
Code smells:
* Highly coupled classes
* The getCommands () and createCommand () methods are linked more to the Command class

Affected class:
* ParamModel.java

Affected methods:
* getCommands()
* createCommand()

Solutions:
* Moving methods related to Command in ParamModel
